x_experiment/pyproject.toml
[project]
name = "socitwin-x-experiment"
version = "0.1.0"
description = "Socitwin Contrarian Agent — X/Twitter 对抗性观点生成与发布"
requires-python = ">=3.11,<3.12"
dependencies = [
    "openai>=1.0.0",
    "tweepy>=4.14.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
